requirements.txt
asgiref==3.5.2
config==0.5.1
Django==4.1.4
psycopg2-binary==2.9.5
python-dotenv==0.21.0
sqlparse==0.4.3

plotly~=5.11.0